requirements.txt
# tkinter is included with Python installation
Pillow>=9.0.0